refactor(vex): split src/vex.rs by format into module directory
Closes #31.

The monolithic src/vex.rs (1441 LOC, 50KB) mixed five concerns:
loader entry point, OpenVEX parser, CycloneDX VEX parser, apply pass,
emit pass, and the synthetic-id formatters/parser. This made the file
the largest module in the crate and well over the soft budget.

Split into a directory module with one file per concern:

  src/vex/mod.rs          — types, load(), detect_format, VexIndex,
                            VexAnnotation, re-exports, loader tests
  src/vex/openvex.rs      — OpenVEX 0.2.0 parser
  src/vex/cyclonedx_vex.rs — CycloneDX VEX 1.6 parser
  src/vex/apply.rs        — apply() pass over enrichment
  src/vex/emit.rs         — emit() and EmitOptions, emission tests
  src/vex/synthetic_id.rs — formatters + parse_synthetic_id, id tests

Public API preserved via re-exports: crate::vex::{load, apply, emit,
EmitOptions, VexAnnotation, VexIndex, VexStatement, VexStatus,
SyntheticFindingKind, parse_synthetic_id, synthetic_id}. All 443 tests
pass.
